requirements.txt
pdfplumber
nltk
sentence-transformers
faiss-cpu
openai
